shift generatesauceurl to a background job
